Python_Bootcamp.Day_06-1/src/requirements.txt
annotated-types==0.6.0
greenlet==3.0.3
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.1
pydantic==2.5.3
pydantic_core==2.14.6
pydentic==0.0.1.dev3
python-stdnum==1.19
SQLAlchemy==2.0.25
typing_extensions==4.9.0
